refactor(ledger): extract skiplist/negativeUNL packages, split service.go
Addresses issue #411 (ledger items):
- #15 move skip-list ops to internal/ledger/skiplist and the NegativeUNL
  flag-ledger transition to internal/ledger/negativeunl; ledger.go 1197->884
- #17 collapse the duplicate amendments LedgerReader onto ledger.Reader
- #18 split service.go (2957->1273) into lifecycle.go/events.go/cache.go
- #19 factor the repeated account-query preamble into withAccountQuery
- #33 trim restating field/interface comments in ledger.go
- #35 cover Insert/Update/Erase error paths, immutability and snapshot
  isolation; fix Erase to return ErrEntryNotFound (was leaking shamap
  ErrItemNotFound, asymmetric with Update)
- #22 unexport genesis serializeAccountRoot for visibility consistency
- #31 unexport manager ledgerCacheConfig/cacheStats
- #20 document the openledger Modify blocking contract

Defers #21 (split state/ into subpackages): 2562 call sites across 227
files makes it an unreviewable sweep, tracked separately on the issue.
